% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment_lynx.R
\name{get_lynx_enrichment}
\alias{get_lynx_enrichment}
\title{Enrichment using Lynx API}
\usage{
get_lynx_enrichment(
  categories = c("go_bp"),
  genelist = c("AKT1", "CASK", "PXN"),
  tax_id = "9606"
)
}
\arguments{
\item{categories}{list of databases}

\item{genelist}{list of genes}

\item{tax_id}{taxonomy id, currently supports human: 9606 and mouse: 10040}
}
\value{
A dataframe of enriched functional categories with pValue, overlapped genes,
corrected pValue, and bayes factor
}
\description{
This function returns the enrichment results for given list of genes
for requested functional categories: disease, pathway, go_bp, go_cc, go_mf
}
