% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment_lynx.R
\name{seurat_marker_enrich}
\alias{seurat_marker_enrich}
\title{Enrichment for single cell cluster markers}
\usage{
seurat_marker_enrich(markers, name, tax_id = "9606")
}
\arguments{
\item{markers}{output from Seurat function, FindAllMarkers}

\item{name}{prefix to save the enrichment results}

\item{tax_id}{taxonomy id, currently supports human: 9606 and mouse: 10040}
}
\value{

}
\description{
This function returns the enrichment results for a single cell marker list
for pre-defined functional categories: pathway, go_bp, go_cc, go_mf
}
